Convert AllAnime to the new parser class
